requirements.txt
python-dotenv
fastapi
uvicorn
rich
requests
pillow
replicate
boto3~=1.28.7
botocore~=1.31.7
openai
streamlit
aiohttp
moviepy

openai-whisper
opencv-python
numpy==1.24.3